requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
crossplane==0.5.8
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
docker==7.1.0
idna==3.10
requests==2.31.0
sqlparse==0.5.3
urllib3==2.3.0
gunicorn==23.0.0
